+++ js/main.js
@@ -3,8 +3,8 @@ import { initPictureModal } from './picture-handler.js';
 import { initUploadPictureForm } from './upload-picture-form.js';
 import { initUploadPictureModal } from './upload-picture-modal.js';
 import { initDataErrorToast } from './data-fetch-error-toast.js';
-import { fetchData } from './utils.js';
-import { BASE_API } from './constants.js';
+import { fetchData } from './shared/fetch.js';
+import { BASE_API } from './shared/constants.js';
 
 const pictureTemplateElement = document
   .querySelector('#picture')
